requirements.txt
Django==6.0
django-allauth==65.13.1
pillow==12.0.0